@@ -1,12 +1,15 @@
 apply plugin: 'com.android.library'
-apply plugin: 'kotlin-android'
+apply plugin: 'org.jetbrains.kotlin.android'
+apply plugin: 'org.jetbrains.kotlin.plugin.compose'
 apply plugin: 'de.undercouch.download'
 apply plugin: 'maven-publish'
 
-android {
-    compileSdkVersion 34
-    defaultConfig {
 
+android {
+  //  namespace "com.assentify.sdk"
+    compileSdkVersion 36
+    defaultConfig {
+    //  applicationId "com.assentify.sdk"
 
         minSdkVersion 24
         targetSdkVersion 34
