@@ -1,6 +1,7 @@
 
 buildscript {
-    ext.kotlin_version = '1.7.0'
+    ext.kotlin_version = '2.2.21'
+
 
     repositories {
         google()
@@ -10,9 +11,11 @@ buildscript {
 
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:7.2.2'
-        classpath 'de.undercouch:gradle-download-task:4.1.1'
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.9.0"
+        classpath 'com.android.tools.build:gradle:8.13.1'
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
+        classpath 'de.undercouch:gradle-download-task:5.6.0'
+        classpath "org.jetbrains.kotlin.plugin.compose:org.jetbrains.kotlin.plugin.compose.gradle.plugin:$kotlin_version"
+
 
     }
 
@@ -22,8 +25,8 @@ buildscript {
 
 allprojects {
     repositories {
-        google()
-        mavenCentral()
+      // google()
+      // mavenCentral()
     }//etc
 }
 
