@@ -1,12 +1,16 @@
 package  com.assentify.sdk
 
 import android.content.Context
+import android.content.Intent
 import android.os.Build
 import android.util.Log
+import com.assentify.sdk.AssistedDataEntry.AssistedDataEntry
+import com.assentify.sdk.AssistedDataEntry.AssistedDataEntryCallback
 import com.assentify.sdk.CheckEnvironment.ContextAwareSigning
 import com.assentify.sdk.ContextAware.ContextAwareSigningCallback
 import com.assentify.sdk.Core.Constants.BlockLoaderKeys
 import com.assentify.sdk.Core.Constants.EnvironmentalConditions
+import com.assentify.sdk.Core.Constants.FlowEnvironmentalConditions
 import com.assentify.sdk.Core.Constants.Language
 import com.assentify.sdk.Core.Constants.StepsNames
 import com.assentify.sdk.Core.Constants.WrapUpKeys
@@ -17,6 +21,8 @@ import com.assentify.sdk.FaceMatch.FaceMatch
 import com.assentify.sdk.FaceMatch.FaceMatchCallback
 import com.assentify.sdk.FaceMatch.FaceMatchManual
 import com.assentify.sdk.FaceMatch.FaceMatchResult
+import com.assentify.sdk.Flow.BlockLoader.BlockLoaderStepsComposeActivity
+import com.assentify.sdk.Flow.Models.FlowCallBack
 import com.assentify.sdk.LanguageTransformation.LanguageTransformation
 import com.assentify.sdk.LanguageTransformation.LanguageTransformationCallback
 import com.assentify.sdk.LanguageTransformation.Models.LanguageTransformationModel
@@ -49,8 +55,6 @@ import com.assentify.sdk.ScanQr.ScanQrManual
 import com.assentify.sdk.ScanQr.ScanQrResult
 import com.assentify.sdk.SubmitData.SubmitData
 import com.assentify.sdk.SubmitData.SubmitDataCallback
-import kotlinx.coroutines.GlobalScope
-import kotlinx.coroutines.launch
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
@@ -68,8 +72,7 @@ class AssentifySdk(
     private var isKeyValid: Boolean = false;
     private var timeStarted: String = "";
     private var configModel: ConfigModel? = null;
-    private var stepID: Int = -1;
-    private var templates: List<TemplatesByCountry> = emptyList();
+    private var allTemplates: List<TemplatesByCountry> = emptyList();
     private lateinit var readJSONFromAsset: ReadJSONFromAsset;
 
 
@@ -140,16 +143,8 @@ class AssentifySdk(
 
 
     private fun iniSdk() {
-        getTemplatesByCountry();
         timeStarted = getCurrentDateTime();
-        GlobalScope.launch {
-            configModel!!.stepDefinitions.forEach { item ->
-                if (item.stepDefinition == "ContextAwareSigning") {
-                    stepID = item.stepId;
-                }
-            }
-
-        }
+        getTemplatesByCountry();
     }
 
     fun startScanPassport(
@@ -158,7 +153,7 @@ class AssentifySdk(
         stepId: Int? = null,
     ): ScanPassportResult {
         if (isKeyValid) {
-            if (ImageUtils.isLowCapabilities(context, this.environmentalConditions)) {
+            if (ImageUtils.isLowCapabilities(context,this.environmentalConditions)) {
                 val scanPassportManual = ScanPassportManual(
                     configModel,
                     environmentalConditions,
@@ -188,11 +183,12 @@ class AssentifySdk(
 
     fun startScanIDCard(
         scnIDCardCallback: IDCardCallback,
-        kycDocumentDetails: List<KycDocumentDetails>, language: String = Language.NON,
+        kycDocumentDetails: List<KycDocumentDetails>,
+        language: String = Language.NON,
         stepId: Int? = null,
     ): ScanIDCardResult {
         if (isKeyValid) {
-            if (ImageUtils.isLowCapabilities(context, this.environmentalConditions)) {
+            if (ImageUtils.isLowCapabilities(context,this.environmentalConditions)) {
                 val scanIDCardManual = ScanIDCardManual(
                     configModel,
                     environmentalConditions, apiKey,
@@ -225,8 +221,8 @@ class AssentifySdk(
         stepId: Int? = null,
     ): ScanQrResult {
         if (isKeyValid) {
-            if (ImageUtils.isLowCapabilities(context, this.environmentalConditions)) {
-                val scanQrManual = ScanQrManual(
+            if (ImageUtils.isLowCapabilities(context,this.environmentalConditions)) {
+                val scanQrManual =  ScanQrManual(
                     kycDocumentDetails,
                     apiKey,
                     language,
@@ -237,7 +233,7 @@ class AssentifySdk(
                 scanQrManual.setScanQrCallback(scanQrCallback)
                 return ScanQrResult.Manual(scanQrManual);
             } else {
-                val scanQr = ScanQr(
+                val scanQr =  ScanQr(
                     kycDocumentDetails,
                     apiKey,
                     language,
@@ -261,7 +257,7 @@ class AssentifySdk(
         stepId: Int? = null,
     ): ScanOtherResult {
         if (isKeyValid) {
-            if (ImageUtils.isLowCapabilities(context, this.environmentalConditions)) {
+            if (ImageUtils.isLowCapabilities(context,this.environmentalConditions)) {
                 val scanOtherManual = ScanOtherManual(
                     configModel,
                     environmentalConditions, apiKey,
@@ -293,7 +289,7 @@ class AssentifySdk(
         stepId: Int? = null,
     ): FaceMatchResult {
         if (isKeyValid) {
-            if (ImageUtils.isLowCapabilities(context, this.environmentalConditions)) {
+            if (ImageUtils.isLowCapabilities(context,this.environmentalConditions)) {
                 val faceMatchManual = FaceMatchManual(
                     configModel,
                     environmentalConditions,
@@ -321,16 +317,30 @@ class AssentifySdk(
         }
     }
 
-    fun startContextAwareSigning(contextAwareSigningCallback: ContextAwareSigningCallback): ContextAwareSigning {
+    fun startContextAwareSigning(contextAwareSigningCallback: ContextAwareSigningCallback,stepId: Int? = null,): ContextAwareSigning {
         if (isKeyValid) {
             return ContextAwareSigning(
                 contextAwareSigningCallback,
                 tenantIdentifier!!,
                 interaction!!,
-                stepID,
+                stepId!!,
                 configModel!!,
                 apiKey!!
             )
+        } else {
+            throw Exception("Invalid Keys")
+        }
+    }
+
+    fun startAssistedDataEntry(assistedDataEntryCallback: AssistedDataEntryCallback,stepId: Int? = null,): AssistedDataEntry {
+        if (isKeyValid) {
+            val assistedDataEntry =  AssistedDataEntry(
+                apiKey!!,
+                configModel!!
+            )
+            assistedDataEntry.setStepId(stepId?.toString())
+            assistedDataEntry.setCallback(assistedDataEntryCallback)
+            return  assistedDataEntry;
         } else {
             throw Exception("Invalid Keys")
         }
@@ -349,6 +359,7 @@ class AssentifySdk(
             appConfiguration = configModel!!,
         )
     }
+
 
 
     fun startSubmitData(
@@ -383,7 +394,6 @@ class AssentifySdk(
             }
             /** BlockLoader **/
             val valuesBlockLoader: MutableMap<String, String> = mutableMapOf()
-
             initSteps.forEach { item ->
                 if (item.stepDefinition == StepsNames.BlockLoader) {
                     item.outputProperties.forEach { property ->
@@ -391,7 +401,10 @@ class AssentifySdk(
                             valuesBlockLoader.put(property.key, timeStarted)
                         }
                         if (property.key.contains(BlockLoaderKeys.DeviceName)) {
-                            valuesBlockLoader.put(property.key, "${Build.MANUFACTURER} ${Build.MODEL}")
+                            valuesBlockLoader.put(
+                                property.key,
+                                "${Build.MANUFACTURER} ${Build.MODEL}"
+                            )
                         }
                         if (property.key.contains(BlockLoaderKeys.Application)) {
                             valuesBlockLoader.put(property.key, configModel!!.applicationId)
@@ -417,7 +430,6 @@ class AssentifySdk(
                                 valuesBlockLoader.put(property.key, customProperty.value.toString())
                             }
                         }
-
                     }
                 }
             }
@@ -431,7 +443,6 @@ class AssentifySdk(
                 ));
             }
             /****/
-
             return SubmitData(apiKey!!, submitDataCallback, submitList, configModel!!)
         } else {
             throw Exception("Invalid Keys")
@@ -463,10 +474,7 @@ class AssentifySdk(
                     }
 
 
-                    templates = filterToSupportedCountries(
-                        templatesByCountry
-                    )!!;
-
+                    allTemplates = templatesByCountry;
                     assentifySdkCallback!!.onAssentifySdkInitSuccess(configModel!!);
                 }
             }
@@ -502,11 +510,11 @@ class AssentifySdk(
     }
 
 
-    private fun filterToSupportedCountries(dataList: List<TemplatesByCountry>?): List<TemplatesByCountry>? {
+    private fun filterToSupportedCountries(dataList: List<TemplatesByCountry>?,stepID:Int): List<TemplatesByCountry>? {
         var selectedCountries: List<String> = emptyList();
         var supportedIdCards: List<String> = emptyList();
         configModel!!.stepDefinitions.forEach { step ->
-            if (step.stepDefinition == "IdentificationDocumentCapture") {
+            if (step.stepId == stepID) {
                 step.customization.identificationDocuments!!.forEach { docStep ->
                     if (docStep.key == "IdentificationDocument.IdCard") {
                         if (docStep.selectedCountries != null) {
@@ -571,7 +579,21 @@ class AssentifySdk(
     }
 
 
-    fun getTemplates() = templates;
+    fun getTemplates(stepID: Int): List<TemplatesByCountry> {
+        val stepTemplates = filterToSupportedCountries(allTemplates, stepID)
+        return stepTemplates ?: emptyList()
+    }
+    /** FLOW **/
 
+    // 1 _ Text && Icons are white
+
+    public fun startFlow(flowCallback: FlowCallBack, flowEnvironmentalConditions: FlowEnvironmentalConditions){
+         FlowEnvironmentalConditionsObject.setFlowEnvironmentalConditions(flowEnvironmentalConditions);
+         ConfigModelObject.setConfigModelObject(configModel!!);
+         FlowCallbackObject.setFlowCallbackObject(flowCallback!!);
+         ApiKeyObject.setApiKeyObject(apiKey!!);
+         val intent = Intent(context, BlockLoaderStepsComposeActivity::class.java)
+         context.startActivity(intent)
+    }
 
 }
