@@ -1,6 +1,6 @@
 package com.assentify.sdk.Core.Constants.Routes
 
 object BaseUrls {
-    const val SignalRHub = "https://widgets.socket.assentify.com/";
+    const val SignalRHub = "https://qa.widgets.socket.assentify.com/";
     const val ContainerName = "orchestrator";
 }
