@@ -9,6 +9,7 @@ import android.graphics.RectF;
 import android.os.Build;
 import android.os.Bundle;
 import android.os.CountDownTimer;
+import android.util.Log;
 import android.view.View;
 
 import androidx.annotation.NonNull;
@@ -172,26 +173,20 @@ public class FaceMatchManual extends CameraPreview implements RemoteProcessingCa
                 start = false;
                 setRectFCustomColor(environmentalConditions.getHoldHandColor(), environmentalConditions.getEnableDetect(), environmentalConditions.getEnableGuide(), start);
                 createBase64.execute(() -> {
-                    remoteProcessing.starProcessing(
+                    remoteProcessing.starProcessingFace(
                             HubConnectionFunctions.INSTANCE.etHubConnectionFunction(BlockType.FACE_MATCH),
-                            "",
-                            ImageUtils.convertBitmapToBase64(normalImage, BlockType.FACE_MATCH, getActivity()),
                             configModel,
-                            "",
-                            this.secondImage,
-                            "ConnectionId",
-                            getVideoPath(configModel, faceMatch, 0),
-                            hasFace(),
-                            true,
-                            true,
-                            performPassiveLivenessFace,
-                            saveCapturedVideo,
-                            true,
-                            true,
-                            storeImageStream,
                             stepId,
-                            new ArrayList<>()
+                            ImageUtils.convertBitmapToByteArray(normalImage, BlockType.FACE_MATCH, getActivity()),
+                            new ArrayList<>(),
+                            ImageUtils.base64ToByteArray(this.secondImage),
+                            performPassiveLivenessFace,
+                            livnessRetryCount,
+                            false,
+                            true,
+                            "ConnectionId"
                     );
+
                 });
             }else {
                 if (getActivity() != null) {
@@ -378,5 +373,11 @@ public class FaceMatchManual extends CameraPreview implements RemoteProcessingCa
         if (this.createBase64 != null) {
             this.createBase64.shutdown();
         }
+    }
+
+    @Override
+    public void onUploadProgress(int progress) {
+        faceMatchCallback.onUploadingProgress(progress);
+
     }
 }
