@@ -25,9 +25,16 @@ fun parseDataToBaseResponseDataModel(data: String): BaseResponseDataModel {
     return BaseResponseDataModel(
         destinationEndpoint = json.optString("destinationEndpoint", ""),
         response = json.optString("response", ""),
-        error = json.optString("error", null),
+        error = json.optString("error", ),
         success = json.optBoolean("success", false)
     )
 }
+
+fun getImageUrlFromBaseResponseDataModel(jsonString: String): String {
+    val json = JSONObject(jsonString)
+    val imageUrl = json.optString("ImageUrl", "")
+    return  imageUrl;
+}
+
 
 
