@@ -1,23 +1,19 @@
 package com.assentify.sdk.Core.Constants
 
-import android.util.Log
-import com.assentify.sdk.Core.Constants.Routes.BaseUrls
+import android.util.Base64
 import com.assentify.sdk.Models.BaseResponseDataModel
 import com.assentify.sdk.Models.parseDataToBaseResponseDataModel
+import com.assentify.sdk.ProcessingRHub.ProgressRequestBody
 import com.assentify.sdk.ProcessingRHub.RemoteProcessingCallback
 import com.assentify.sdk.RemoteClient.Models.ConfigModel
 import com.assentify.sdk.RemoteClient.RemoteClient
 import okhttp3.MediaType.Companion.toMediaTypeOrNull
 import okhttp3.MultipartBody
-import okhttp3.RequestBody
+import okhttp3.RequestBody.Companion.toRequestBody
 import okhttp3.ResponseBody
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
-import retrofit2.http.Header
-import retrofit2.http.Part
-import retrofit2.http.Url
-import java.io.File
 import java.util.UUID
 
 class RemoteProcessing {
@@ -28,166 +24,19 @@ class RemoteProcessing {
         this.callback = callback
     }
 
-    fun uploadVideo(
-        counter: Int,
-        uploadFile: File,
-        appConfiguration: ConfigModel,
-        templateId: String
-    ) {
-        val fileRequestBody = RequestBody.create(null, File(uploadFile.absolutePath))
-        val filePart = MultipartBody.Part.createFormData(
-            "files", "${counter}.mp4", fileRequestBody
-        )
-        val call = RemoteClient.remoteBlobStorageService.uploadFile(
-            BaseUrls.ContainerName,
-            "${counter}",
-            filePart,
-            RequestBody.create("text/plain".toMediaTypeOrNull(), appConfiguration.tenantIdentifier),
-            RequestBody.create("text/plain".toMediaTypeOrNull(), appConfiguration.blockIdentifier),
-            RequestBody.create("text/plain".toMediaTypeOrNull(), appConfiguration.instanceId),
-            RequestBody.create("text/plain".toMediaTypeOrNull(), templateId),
-            RequestBody.create("text/plain".toMediaTypeOrNull(), "try_${counter}"),
-        )
-        call.enqueue(object : Callback<ResponseBody> {
-            override fun onResponse(
-                call: Call<ResponseBody>,
-                response: Response<ResponseBody>
-            ) {
-                if (response.isSuccessful) {
-                    val responseBody = response.body()
-                    if (responseBody != null) {
-                        val responseBodyString = responseBody.string()
-
-                    }
-                }
-            }
-
-            override fun onFailure(call: Call<ResponseBody>, t: Throwable) {
-            }
-        })
-
-    }
-
-    fun starProcessing(
-        url: String,
-        videoClip: String,
-        selfieImage: String,
-        appConfiguration: ConfigModel,
-        templateId: String,
-        secondImage: String,
-        connectionId: String,
-        clipsPath: String,
-        checkForFace: Boolean,
-        processMrz: Boolean,
-        performLivenessDoc: Boolean,
-        performLivenessFace: Boolean,
-        saveCapturedVideo: Boolean,
-        storeCapturedDocument: Boolean,
-        isVideo: Boolean,
-        storeImageStream: Boolean,
-        stepId: String,
-        clipParts: List<String>,
-    ) {
-        val traceIdentifier = UUID.randomUUID().toString()
-
-        val clips = clipParts.mapIndexed { index, clip ->
-            MultipartBody.Part.createFormData(
-                "clips[$index]",
-                clip
-            )
-        }
-
-
-        val call = RemoteClient.remoteWidgetsService.starProcessing(
-            url,
-            stepId,
-            appConfiguration.blockIdentifier,
-            appConfiguration.flowIdentifier,
-            appConfiguration.flowInstanceId,
-            appConfiguration.instanceHash,
-            appConfiguration.instanceId,
-            appConfiguration.tenantIdentifier,
-            RequestBody.create("text/plain".toMediaTypeOrNull(), appConfiguration.tenantIdentifier),
-            RequestBody.create("text/plain".toMediaTypeOrNull(), appConfiguration.blockIdentifier),
-            RequestBody.create("text/plain".toMediaTypeOrNull(), appConfiguration.instanceId),
-            RequestBody.create("text/plain".toMediaTypeOrNull(), templateId),
-            RequestBody.create(
-                "text/plain".toMediaTypeOrNull(),
-                performLivenessDoc.toString()
-            ),
-            RequestBody.create(
-                "text/plain".toMediaTypeOrNull(),
-                performLivenessFace.toString()
-            ),
-            RequestBody.create("text/plain".toMediaTypeOrNull(), processMrz.toString()),
-            RequestBody.create("text/plain".toMediaTypeOrNull(), "false"),
-            RequestBody.create("text/plain".toMediaTypeOrNull(), storeImageStream.toString()),
-            RequestBody.create("text/plain".toMediaTypeOrNull(), isVideo.toString()),
-            RequestBody.create("text/plain".toMediaTypeOrNull(), clipsPath),
-            RequestBody.create("text/plain".toMediaTypeOrNull(), "true"),
-            RequestBody.create("text/plain".toMediaTypeOrNull(), videoClip),
-            RequestBody.create("text/plain".toMediaTypeOrNull(), checkForFace.toString()),
-            RequestBody.create("text/plain".toMediaTypeOrNull(), connectionId),
-            RequestBody.create("text/plain".toMediaTypeOrNull(), secondImage),
-            RequestBody.create("text/plain".toMediaTypeOrNull(), saveCapturedVideo.toString()),
-            RequestBody.create("text/plain".toMediaTypeOrNull(), storeCapturedDocument.toString()),
-            RequestBody.create("text/plain".toMediaTypeOrNull(), traceIdentifier),
-            RequestBody.create("text/plain".toMediaTypeOrNull(), selfieImage),
-            clips,
-            )
-        call.enqueue(object : Callback<ResponseBody> {
-            override fun onResponse(
-                call: Call<ResponseBody>,
-                response: Response<ResponseBody>
-            ) {
-
-                if (response.isSuccessful) {
-                    val responseBody = response.body()?.string() ?: ""
-                    val baseResponseDataModel = parseDataToBaseResponseDataModel(responseBody)
-                    callback!!.onMessageReceived(
-                        baseResponseDataModel.destinationEndpoint!!,
-                        baseResponseDataModel
-                    )
-
-                }else {
-                    val responseBody = response.body()?.string() ?: ""
-                    callback!!.onMessageReceived(
-                        HubConnectionTargets.ON_ERROR,
-                        BaseResponseDataModel(
-                            destinationEndpoint = HubConnectionTargets.ON_ERROR,
-                            response = "",
-                            error = "",
-                            success = false
-                        )
-                    );
-                }
-            }
-
-            override fun onFailure(call: Call<ResponseBody>, t: Throwable) {
-                callback!!.onMessageReceived(
-                    HubConnectionTargets.ON_ERROR,
-                    BaseResponseDataModel(
-                        destinationEndpoint = HubConnectionTargets.ON_ERROR,
-                        response = "",
-                        error = t.message,
-                        success = false
-                    )
-                );
-            }
-        })
-
-
-    }
 
     fun starQrProcessing(
         url: String,
-        videoClip: String,
+        byteArrayImage: ByteArray,
         appConfiguration: ConfigModel,
         templateId: String,
         connectionId: String,
         stepId: String,
         metadata: String,
+        isManualCapture : Boolean,
+        isAutoCapture : Boolean,
     ) {
+        val traceIdentifier = UUID.randomUUID().toString()
         val call = RemoteClient.remoteWidgetsService.starQrProcessing(
             url,
             stepId,
@@ -197,17 +46,26 @@ class RemoteProcessing {
             appConfiguration.instanceHash,
             appConfiguration.instanceId,
             appConfiguration.tenantIdentifier,
-            RequestBody.create("text/plain".toMediaTypeOrNull(), appConfiguration.tenantIdentifier),
-            RequestBody.create("text/plain".toMediaTypeOrNull(), appConfiguration.blockIdentifier),
-            RequestBody.create("text/plain".toMediaTypeOrNull(), appConfiguration.instanceId),
-            RequestBody.create("text/plain".toMediaTypeOrNull(), templateId),
-            RequestBody.create(
-                "text/plain".toMediaTypeOrNull(),
-                "true"
-            ),
-            RequestBody.create("text/plain".toMediaTypeOrNull(), videoClip),
-            RequestBody.create("text/plain".toMediaTypeOrNull(), connectionId),
-            RequestBody.create("text/plain".toMediaTypeOrNull(), metadata),
+            appConfiguration.tenantIdentifier.toRequestBody("text/plain".toMediaTypeOrNull()),
+            appConfiguration.blockIdentifier.toRequestBody("text/plain".toMediaTypeOrNull()),
+            appConfiguration.instanceId.toRequestBody("text/plain".toMediaTypeOrNull()),
+            templateId.toRequestBody("text/plain".toMediaTypeOrNull()),
+            "true"
+                .toRequestBody("text/plain".toMediaTypeOrNull()),
+            byteArrayToPart(
+                byteArray = byteArrayImage,
+                partName = "Image",
+                fileName = "image.jpg",
+                mimeType = "image/jpeg"
+            ){ sent, total, done ->
+                val pct = if (total > 0) ((sent * 100) / total).toInt() else -1
+                callback!!.onUploadProgress(pct);
+            },
+            connectionId.toRequestBody("text/plain".toMediaTypeOrNull()),
+            metadata.toRequestBody("text/plain".toMediaTypeOrNull()),
+            traceIdentifier.toRequestBody("text/plain".toMediaTypeOrNull()),
+            isManualCapture.toString().toRequestBody("text/plain".toMediaTypeOrNull()),
+            isAutoCapture.toString().toRequestBody("text/plain".toMediaTypeOrNull()),
         )
         call.enqueue(object : Callback<ResponseBody> {
             override fun onResponse(
@@ -253,6 +111,269 @@ class RemoteProcessing {
 
     }
 
+    fun starProcessingIDs(
+        url: String,
+        byteArrayImage: ByteArray,
+        appConfiguration: ConfigModel,
+        templateId: String,
+        connectionId: String,
+        clipsPath: String,
+        checkForFace: Boolean,
+        processMrz: Boolean,
+        performLivenessDoc: Boolean,
+        saveCapturedVideo: Boolean,
+        storeCapturedDocument: Boolean,
+        storeImageStream: Boolean,
+        stepId: String,
+        isManualCapture : Boolean,
+        isAutoCapture : Boolean,
+        tryNumber:Int,
+        tag:String,
+        processCivilExtractQrCode:Boolean,
+    ) {
+        val tryCount  = tryNumber + 1;
+        val traceIdentifier = UUID.randomUUID().toString()
+        val call = RemoteClient.remoteWidgetsService.starProcessingIDs(
+            url,
+            stepId,
+            appConfiguration.blockIdentifier,
+            appConfiguration.flowIdentifier,
+            appConfiguration.flowInstanceId,
+            appConfiguration.instanceHash,
+            appConfiguration.instanceId,
+            appConfiguration.tenantIdentifier,
+            appConfiguration.tenantIdentifier.toRequestBody("text/plain".toMediaTypeOrNull()),
+            appConfiguration.blockIdentifier.toRequestBody("text/plain".toMediaTypeOrNull()),
+            appConfiguration.instanceId.toRequestBody("text/plain".toMediaTypeOrNull()),
+            templateId.toRequestBody("text/plain".toMediaTypeOrNull()),
+            performLivenessDoc.toString()
+                .toRequestBody("text/plain".toMediaTypeOrNull()),
+            processMrz.toString().toRequestBody("text/plain".toMediaTypeOrNull()),
+            "false".toRequestBody("text/plain".toMediaTypeOrNull()), //  DisableDataExtraction
+            storeImageStream.toString().toRequestBody("text/plain".toMediaTypeOrNull()),
+            "false".toRequestBody("text/plain".toMediaTypeOrNull()), // isVideo
+            clipsPath.toRequestBody("text/plain".toMediaTypeOrNull()),
+            "true".toRequestBody("text/plain".toMediaTypeOrNull()), // isMobile
+            byteArrayToPart(
+                byteArray = byteArrayImage,
+                partName = "Image",
+                fileName = "image.jpg",
+                mimeType = "image/jpeg"
+            ){ sent, total, done ->
+                val pct = if (total > 0) ((sent * 100) / total).toInt() else -1
+                callback!!.onUploadProgress(pct);
+            },
+            checkForFace.toString().toRequestBody("text/plain".toMediaTypeOrNull()),
+            connectionId.toRequestBody("text/plain".toMediaTypeOrNull()),
+            saveCapturedVideo.toString().toRequestBody("text/plain".toMediaTypeOrNull()),
+            storeCapturedDocument.toString().toRequestBody("text/plain".toMediaTypeOrNull()),
+            traceIdentifier.toRequestBody("text/plain".toMediaTypeOrNull()),
+            isManualCapture.toString().toRequestBody("text/plain".toMediaTypeOrNull()),
+            isAutoCapture.toString().toRequestBody("text/plain".toMediaTypeOrNull()),
+            tryCount.toString().toRequestBody("text/plain".toMediaTypeOrNull()),
+            tag.toRequestBody("text/plain".toMediaTypeOrNull()),
+            processCivilExtractQrCode.toString().toRequestBody("text/plain".toMediaTypeOrNull()),
+            "false".toRequestBody("text/plain".toMediaTypeOrNull()), // RequireFaceExtraction
+        )
+        call.enqueue(object : Callback<ResponseBody> {
+            override fun onResponse(
+                call: Call<ResponseBody>,
+                response: Response<ResponseBody>
+            ) {
 
+                if (response.isSuccessful) {
+                    val responseBody = response.body()?.string() ?: ""
+                    val baseResponseDataModel = parseDataToBaseResponseDataModel(responseBody)
+                    callback!!.onMessageReceived(
+                        baseResponseDataModel.destinationEndpoint!!,
+                        baseResponseDataModel
+                    )
+
+                }else {
+                    val responseBody = response.body()?.string() ?: ""
+                    callback!!.onMessageReceived(
+                        HubConnectionTargets.ON_ERROR,
+                        BaseResponseDataModel(
+                            destinationEndpoint = HubConnectionTargets.ON_ERROR,
+                            response = "",
+                            error = "",
+                            success = false
+                        )
+                    );
+                }
+            }
+
+            override fun onFailure(call: Call<ResponseBody>, t: Throwable) {
+                callback!!.onMessageReceived(
+                    HubConnectionTargets.ON_ERROR,
+                    BaseResponseDataModel(
+                        destinationEndpoint = HubConnectionTargets.ON_ERROR,
+                        response = "",
+                        error = t.message,
+                        success = false
+                    )
+                );
+            }
+        })
+
+
+    }
+
+    fun starProcessingFace(
+        url: String,
+        appConfiguration: ConfigModel,
+        stepId: String,
+        selfieImage: ByteArray,
+        livenessFrames:  List<ByteArray>,
+        secondImage: ByteArray,
+        isLivenessEnabled:Boolean,
+        tryNumber:Int,
+        isAutoCapture : Boolean,
+        isManualCapture : Boolean,
+        connectionId: String,
+    ) {
+        var currentClip = 0;
+        val tryCount  = tryNumber + 1;
+        val traceIdentifier = UUID.randomUUID().toString()
+        val call = RemoteClient.remoteWidgetsService.starProcessingFace(
+            url,
+            stepId,
+            appConfiguration.blockIdentifier,
+            appConfiguration.flowIdentifier,
+            appConfiguration.flowInstanceId,
+            appConfiguration.instanceHash,
+            appConfiguration.instanceId,
+            appConfiguration.tenantIdentifier,
+            appConfiguration.tenantIdentifier.toRequestBody("text/plain".toMediaTypeOrNull()),
+            appConfiguration.blockIdentifier.toRequestBody("text/plain".toMediaTypeOrNull()),
+            appConfiguration.instanceId.toRequestBody("text/plain".toMediaTypeOrNull()),
+            byteArrayToPart(
+                byteArray = selfieImage,
+                partName = "SelfieImage",
+                fileName = "selfieImage.jpg",
+                mimeType = "image/jpeg"
+            ){ sent, total, done ->
+                val pct = if (total > 0) ((sent * 100) / total).toInt() else -1
+                if(isManualCapture){
+                        callback!!.onUploadProgress(pct);
+                }
+                if(isAutoCapture){
+                    if(!isLivenessEnabled){
+                        callback!!.onUploadProgress(pct);
+
+                    }
+                }
+            },
+            byteArraysToParts(
+                byteArrays = livenessFrames,
+                partName = "livenessFrames",
+                filePrefix = "frame",
+                mimeType = "image/jpeg",
+            ){ sent, total, done ->
+                val pct = if (total > 0) ((sent * 100) / total).toInt() else -1
+                if(isAutoCapture){
+                    if(isLivenessEnabled){
+                        if(pct == 100){
+                            currentClip += 1;
+                            val currentProgress  = currentClip *  8.33333333
+                            callback!!.onUploadProgress(currentProgress.toInt());
+                        }
+
+                    }
+                }
+            },
+            traceIdentifier.toRequestBody("text/plain".toMediaTypeOrNull()),
+            "true".toRequestBody("text/plain".toMediaTypeOrNull()), // isMobile
+            byteArrayToPart(
+                byteArray = secondImage,
+                partName = "SecondImage",
+                fileName = "selfieImage.jpg",
+                mimeType = "image/jpeg"
+            ){ sent, total, done ->
+              //  val pct = if (total > 0) ((sent * 100) / total).toInt() else -1
+              //  callback!!.onUploadProgress(pct);
+            },
+            isLivenessEnabled.toString().toRequestBody("text/plain".toMediaTypeOrNull()),
+            tryCount.toString().toRequestBody("text/plain".toMediaTypeOrNull()),
+            isAutoCapture.toString().toRequestBody("text/plain".toMediaTypeOrNull()),
+            isManualCapture.toString().toRequestBody("text/plain".toMediaTypeOrNull()),
+            connectionId.toRequestBody("text/plain".toMediaTypeOrNull()),
+
+        )
+        call.enqueue(object : Callback<ResponseBody> {
+            override fun onResponse(
+                call: Call<ResponseBody>,
+                response: Response<ResponseBody>
+            ) {
+
+                if (response.isSuccessful) {
+                    val responseBody = response.body()?.string() ?: ""
+                    val baseResponseDataModel = parseDataToBaseResponseDataModel(responseBody)
+                    callback!!.onMessageReceived(
+                        baseResponseDataModel.destinationEndpoint!!,
+                        baseResponseDataModel
+                    )
+
+                }else {
+                    val responseBody = response.body()?.string() ?: ""
+                    callback!!.onMessageReceived(
+                        HubConnectionTargets.ON_ERROR,
+                        BaseResponseDataModel(
+                            destinationEndpoint = HubConnectionTargets.ON_ERROR,
+                            response = "",
+                            error = "",
+                            success = false
+                        )
+                    );
+                }
+            }
+
+            override fun onFailure(call: Call<ResponseBody>, t: Throwable) {
+                callback!!.onMessageReceived(
+                    HubConnectionTargets.ON_ERROR,
+                    BaseResponseDataModel(
+                        destinationEndpoint = HubConnectionTargets.ON_ERROR,
+                        response = "",
+                        error = t.message,
+                        success = false
+                    )
+                );
+            }
+        })
+
+
+    }
+
+    private fun byteArrayToPart(
+        byteArray: ByteArray,
+        partName: String,
+        fileName: String,
+        mimeType: String,
+        onProgress: (bytesWritten: Long, contentLength: Long, done: Boolean) -> Unit
+    ): MultipartBody.Part {
+        val requestBody = byteArray.toRequestBody(mimeType.toMediaTypeOrNull())
+        val image =  Base64.encodeToString(byteArray, Base64.DEFAULT);
+       // Log.e("byteArrayToPart",image.toString())
+        val progressBody = ProgressRequestBody(requestBody, onProgress)
+        return MultipartBody.Part.createFormData(partName, fileName, progressBody)
+    }
+
+    private fun byteArraysToParts(
+        byteArrays: List<ByteArray>,
+        partName: String,
+        filePrefix: String,
+        mimeType: String,
+        onProgress: (bytesWritten: Long, contentLength: Long, done: Boolean) -> Unit
+    ): List<MultipartBody.Part> {
+        return byteArrays.mapIndexed { index, byteArray ->
+            val fileName = "${filePrefix}_${index}.jpg"
+            val requestBody = byteArray.toRequestBody(mimeType.toMediaTypeOrNull())
+            val image =  Base64.encodeToString(byteArray, Base64.DEFAULT);
+        //    Log.e("byteArrayToPart",image.toString())
+            val progressBody = ProgressRequestBody(requestBody, onProgress)
+            MultipartBody.Part.createFormData(partName, fileName, progressBody)
+        }
+    }
 
 }
+
